% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addresses}
\alias{addresses}
\alias{addresses<-}
\title{User & group addresses}
\usage{
addresses(object)

addresses(object) <- value
}
\arguments{
\item{object}{an instance of TheCity class}

\item{value}{a data frame of addresses}
}
\value{
a data frame of addresses
}
\description{
User & group addresses
}

