% Generated by roxygen2 (4.0.2): do not edit by hand
\name{campuses}
\alias{campuses}
\alias{campuses<-}
\title{Church campuses}
\usage{
campuses(object)

campuses(object) <- value
}
\arguments{
\item{object}{an instance of TheCity class}

\item{value}{a data frame of campuses}
}
\value{
a data frame of campuses
}
\description{
Church campuses
}

