% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkins}
\alias{checkins}
\alias{checkins<-}
\title{Event checkins}
\usage{
checkins(object)

checkins(object) <- value
}
\arguments{
\item{object}{an instance of TheCity class}

\item{value}{a data frame of event checkins}
}
\value{
a data frame of event checkins
}
\description{
Event checkins
}

