% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fetch.addresses}
\alias{fetch.addresses}
\title{Gets addresses via The City API}
\usage{
fetch.addresses(object, start, total)
}
\arguments{
\item{object}{an instance of TheCity class}

\item{start}{the page to start the request on. Default: 1.}

\item{total}{the number of pages of addresses to request. Default: all}
}
\description{
Gets addresses via The City API
}

