% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fetch.checkins}
\alias{fetch.checkins}
\title{Gets event checkins via The City API}
\usage{
fetch.checkins(object, start = 1, total = "all", sleep = TRUE,
  params = c())
}
\arguments{
\item{object}{an instance of TheCity class}

\item{start}{the page to start the request on. Default: 1.}

\item{total}{the number of pages of event checkins to request. Default: all}

\item{sleep}{if TRUE a sleep time will be inserted between page requests.
Default: TRUE}

\item{param}{a named vector of query parameters.}
}
\description{
Gets event checkins via The City API
}

