% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fetch.roles}
\alias{fetch.roles}
\title{Gets user roles via The City API}
\usage{
fetch.roles(object, start = 1, total = "all", sleep = TRUE)
}
\arguments{
\item{object}{an instance of TheCity class}

\item{start}{the page to start the request on. Default: 1.}

\item{total}{the number of pages of roles to request. Default: all}

\item{sleep}{if TRUE a sleep time will be inserted between page requests.
Default: TRUE}
}
\description{
Gets user roles via The City API
}

