% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fetch.skills}
\alias{fetch.skills}
\title{Gets user skills via The City API}
\usage{
fetch.skills(object, ...)
}
\arguments{
\item{object}{an instance of TheCity class}

\item{start}{the page to start the request on. Default: 1.}

\item{total}{the number of pages of skills to request. Default: all}

\item{sleep}{if TRUE a sleep time will be inserted between page requests.
Default: TRUE}
}
\description{
Gets user skills via The City API
}

