% Generated by roxygen2 (4.0.2): do not edit by hand
\name{groups}
\alias{groups}
\alias{groups<-}
\title{Groups}
\usage{
groups(object)

groups(object) <- value
}
\arguments{
\item{object}{an instance of TheCity class}

\item{value}{a data frame of groups}
}
\value{
a data frame of groups
}
\description{
Groups
}

