% Generated by roxygen2 (4.0.2): do not edit by hand
\name{key}
\alias{key}
\alias{key<-}
\title{The Secret Key, used during an API request}
\usage{
key(object)

key(object) <- value
}
\arguments{
\item{object}{an instance of TheCity class}

\item{value}{the new value of the secret key}
}
\value{
the secret key
}
\description{
The Secret Key, used during an API request
}

