% Generated by roxygen2 (4.0.2): do not edit by hand
\name{max.sleep.time}
\alias{max.sleep.time}
\alias{max.sleep.time<-}
\title{Maximum sleep time between API requests}
\usage{
max.sleep.time(object)

max.sleep.time(object) <- value
}
\arguments{
\item{object}{an instance of TheCity class}

\item{value}{the new value of the sleep time in minutes}
}
\value{
the maximum sleep time, in minutes
}
\description{
Maximum sleep time between API requests
}

