% Generated by roxygen2 (4.0.2): do not edit by hand
\name{save.addresses}
\alias{save.addresses}
\title{Saves addresses to an RDS file}
\usage{
save.addresses(object, file)
}
\arguments{
\item{object}{an instance of TheCity class}

\item{file}{the file path to save the data at.}
}
\description{
Saves addresses to an RDS file
}

