% Generated by roxygen2 (4.0.2): do not edit by hand
\name{save.checkins}
\alias{save.checkins}
\title{Saves event checkins to an RDS file}
\usage{
save.checkins(object, file)
}
\arguments{
\item{object}{an instance of TheCity class}

\item{file}{the file path to save the data at}
}
\description{
Saves event checkins to an RDS file
}

