% Generated by roxygen2 (4.0.2): do not edit by hand
\name{save.users}
\alias{save.users}
\title{Saves users to an RDS file}
\usage{
save.users(object, file)
}
\arguments{
\item{object}{an instance of TheCity class}

\item{file}{the file path to save the data at}
}
\description{
Saves users to an RDS file
}

