% Generated by roxygen2 (4.0.2): do not edit by hand
\name{skills}
\alias{skills}
\alias{skills<-}
\title{User skills}
\usage{
skills(object)

skills(object) <- value
}
\arguments{
\item{object}{an instance of TheCity object}

\item{value}{a data frame of skills}
}
\value{
a data frame of skills
}
\description{
User skills
}

