% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{TheCity-class}
\alias{TheCity}
\alias{TheCity-class}
\title{An S4 class that defines an API connection to The City & stores data from The
City}
\description{
An S4 class that defines an API connection to The City & stores data from The
City
}
\section{Slots}{

\describe{
\item{\code{key}}{the Secret Key}

\item{\code{token}}{the User Token}

\item{\code{max.sleep.time}}{the maximum sleep time between API request in minutes. Default 1.25 minutes.}

\item{\code{env}}{the environment where data from The City will be stored.}
}}

