% Generated by roxygen2 (4.0.2): do not edit by hand
\name{token}
\alias{token}
\alias{token<-}
\title{The User Token, used during an API request}
\usage{
token(object)

token(object) <- value
}
\arguments{
\item{object}{an instance of TheCity class}

\item{value}{the new value of the user token}
}
\value{
the user token
}
\description{
The User Token, used during an API request
}

