% Generated by roxygen2 (4.0.2): do not edit by hand
\name{user}
\alias{user}
\title{A user on The City}
\usage{
user(object, usid)
}
\arguments{
\item{object}{an instance of TheCity class}

\item{usid}{the user id of the given user}
}
\value{
a data frame of the user
}
\description{
Looks in TheCity object for the user. If not found, looks to the API & adds
the user to the stored users in TheCity object.
}

