% Generated by roxygen2 (4.0.2): do not edit by hand
\name{users}
\alias{users}
\alias{users<-}
\title{The City users}
\usage{
users(object)

users(object) <- value
}
\arguments{
\item{object}{an instance of TheCity class}

\item{value}{a data frame of users}
}
\value{
a data frame of users
}
\description{
The City users
}

